+++ flow-report/src/app.tsx
@@ -5,11 +5,7 @@
  */
 
 import {FunctionComponent} from 'preact';
-import {useMemo} from 'preact/hooks';
-import {CategoryRenderer} from '../../report/renderer/category-renderer';
-import {DetailsRenderer} from '../../report/renderer/details-renderer';
-import {DOM} from '../../report/renderer/dom';
-import {ReportRendererContext} from './legacy-wrappers';
+import {LegacyRendererWrapper} from './legacy-wrappers';
 import {Sidebar} from './sidebar/sidebar';
 import {Summary} from './summary/summary';
 import {FlowResultContext, useCurrentLhr} from './util';
@@ -34,24 +30,14 @@ const Content: FunctionComponent = () => {
 };
 
 export const App: FunctionComponent<{flowResult: LH.FlowResult}> = ({flowResult}) => {
-  const globals = useMemo(() => {
-    const dom = new DOM(document);
-    const detailsRenderer = new DetailsRenderer(dom);
-    const categoryRenderer = new CategoryRenderer(dom, detailsRenderer);
-    return {
-      dom,
-      detailsRenderer,
-      categoryRenderer,
-    };
-  }, []);
   return (
     <FlowResultContext.Provider value={flowResult}>
-      <ReportRendererContext.Provider value={globals}>
+      <LegacyRendererWrapper>
         <div className="App">
           <Sidebar/>
           <Content/>
         </div>
-      </ReportRendererContext.Provider>
+      </LegacyRendererWrapper>
     </FlowResultContext.Provider>
   );
 };
